+++ src/modules/aws/types/aws.types.ts
@@ -1,0 +1,2 @@
+export type AwsUploadDomain = 'events' | 'community';
+export type AwsUploadFileType = 'image' | 'video';
